Use record count from system
pfgrep currently calculates the record count via size / (length - 12).
The file size is from stat() and the length is the record length,
after subtracting the 6 character date and 6 character sequence.

However, in some situations, it seems this is inaccurate, and pfgrep
ends up reading an incorrect number of records (usually more).

This is the most accurate, but might be expensive. If needed, we can
avoid the member description reading in most cases.